// src/enums/EnvConfigEnum.ts
export enum EnvConfigEnum {
  HOST = 'HOST',
  HOST_PORT = 'HOST_PORT',
  DB_HOST = 'DB_HOST',
  DB_PORT = 'DB_PORT',
  DB_USERNAME = 'DB_USERNAME',
  DB_PASSWORD = 'DB_PASSWORD',
  DB_NAME = 'DB_NAME',
  REDIS_HOST = 'REDIS_HOST',
  REDIS_PORT = 'REDIS_PORT',
  JWT_SECRET = 'JWT_SECRET',
  JWT_EXPIRATION = 'JWT_EXPIRATION',
  STRIPE_SECRET_KEY = 'STRIPE_SECRET_KEY',
  STRIPE_WEBHOOK_SECRET = 'STRIPE_WEBHOOK_SECRET',
  ELASTICSEARCH_NODE = 'ELASTICSEARCH_NODE',
  ELASTICSEARCH_USERNAME = 'ELASTICSEARCH_USERNAME',
  ELASTICSEARCH_PASSWORD = 'ELASTICSEARCH_PASSWORD',
  LOGSTASH_HOST = 'LOGSTASH_HOST',
  LOGSTASH_PORT = 'LOGSTASH_PORT',
  THROTTLE_TTL = 'THROTTLE_TTL',
  THROTTLE_LIMIT = 'THROTTLE_LIMIT',
}